fix(metaengine): float64 planned columns no longer silently become TEXT
BuildLayoutPlanFromType mapped float64 fields to the SQL type name
"DOUBLE", but the pgengine and mysqlengine planned-table DDL translators
only recognize "REAL" and "INTEGER" — a "DOUBLE" column fell through to
their TEXT default, so a registered float64 sort/filter column was created
as TEXT and numeric comparisons degenerated (lexical ordering or typed
failures). DuckDB accepted both names, which is why the divergence went
unnoticed.

The fix maps float64 (and float32) to the canonical "REAL": pgengine
translates REAL to DOUBLE PRECISION, mysqlengine to DOUBLE, duckdb accepts
REAL. PlansColumnCompatible compares column NAMES only, so existing
registrations are unaffected; new plans get numeric columns.

Also: BuildLayoutPlanFromType now indexes fields by their json-tag name as
well as the Go name (same resolution ExtractFields uses), so filters on
the wire name derive the Go field's type.

Proofs: new live tests on PG and MariaDB register a plan from a struct
with a float64 field, then push a numeric filter + sort through the
extracted column and assert numeric ordering (0.25 < 3.75 < 9.5) — a TEXT
column fails this. Unit tests pin the corrected mapping and the json-tag
aliasing. Gates: metaengine lint 0 issues, api golden stable at 4347,
duplicate file suppressed in favor of the existing richer implementation.

+++ metaengine/layout_type_test.go
@@ -29,7 +29,7 @@ func TestBuildLayoutPlanFromType_InfersColumnTypes(t *testing.T) {
 	cases := map[string]string{
 		"status":   "TEXT",    // string
 		"priority": "INTEGER", // int
-		"score":    "DOUBLE",  // float64
+		"score":    "REAL",    // float64 — canonical float name: pg maps REAL→DOUBLE PRECISION, mysql REAL→DOUBLE, duckdb accepts REAL
 		"name":     "TEXT",    // string
 	}
